oneCycle and payment completed XmlReader done BankImpl loadData func Derisk Yaz Pair Storage ID from Integer to String XML Exceptions Main Menu

+++ src/bank/loans/interest/Interest.java
@@ -1,14 +1,14 @@
 package bank.loans.interest;
 
 public interface Interest {
-    float getFinalAmount();
-    float getInterest(); // returns calculated amount of interest to all payments
+    int getFinalAmount();
+    int getInterest(); // returns calculated amount of interest to all payments
 
     int getCyclesPerPayment();
 
     int getDuration();
 
     float getPercent(); // returns interest's percent per payment
-    float getBaseAmount();
+    int getBaseAmount();
 
 }
